Moved Home and Add Word to drawer Moved Fragments directory under ui package Added new items to the drawer
