+++ app/src/main/java/com/teamzmron/selfstudyapp/di/main/MainFragmentsModule.kt
@@ -1,9 +1,9 @@
 package com.teamzmron.selfstudyapp.di.main
 
-import com.teamzmron.selfstudyapp.Fragments.Adjective.AdjectiveHomeFragment
-import com.teamzmron.selfstudyapp.Fragments.Home
-import com.teamzmron.selfstudyapp.Fragments.noun.NounHomeFragment
-import com.teamzmron.selfstudyapp.Fragments.verb.VerbHomeFragment
+import com.teamzmron.selfstudyapp.ui.Fragments.Adjective.AdjectiveHomeFragment
+import com.teamzmron.selfstudyapp.ui.Fragments.Home
+import com.teamzmron.selfstudyapp.ui.Fragments.noun.NounHomeFragment
+import com.teamzmron.selfstudyapp.ui.Fragments.verb.VerbHomeFragment
 
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
