+++ app/src/main/java/com/teamzmron/selfstudyapp/ui/Fragments/Adjective/AdjectiveHomeFragment.kt
@@ -1,22 +1,22 @@
-package com.teamzmron.selfstudyapp.Fragments.verb
+package com.teamzmron.selfstudyapp.ui.Fragments.Adjective
 
 import android.os.Bundle
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.teamzmron.selfstudyapp.Adapters.VerbAdapter
-import com.teamzmron.selfstudyapp.Helper.VerbSwipeToDeleteHelper
+import com.teamzmron.selfstudyapp.Adapters.AdjectiveAdapter
+import com.teamzmron.selfstudyapp.Helper.AdjectiveSwipeToDeleteHelper
+
 import com.teamzmron.selfstudyapp.R
-import com.teamzmron.selfstudyapp.ViewModel.VerbViewModel
+import com.teamzmron.selfstudyapp.ViewModel.AdjectiveViewModel
 import com.teamzmron.selfstudyapp.ViewModel.ViewModelProviderFactory
 import dagger.android.support.DaggerFragment
-import kotlinx.android.synthetic.main.fragment_verb_home.*
+import kotlinx.android.synthetic.main.fragment_adjective_home.*
 import javax.inject.Inject
-
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -25,21 +25,22 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [VerbHomeFragment.newInstance] factory method to
+ * Use the [AdjectiveHomeFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class VerbHomeFragment : DaggerFragment(), VerbAdapter.OnVerbClickListener {
+class AdjectiveHomeFragment : DaggerFragment(), AdjectiveAdapter.OnAdjectiveClickListener {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
 
-    private lateinit var verbViewModel: VerbViewModel
+    private lateinit var adjectiveViewModel: AdjectiveViewModel
 
     @Inject
-    lateinit var verbAdapter: VerbAdapter
+    lateinit var adapter: AdjectiveAdapter
 
     @Inject
     lateinit var providerFactory: ViewModelProviderFactory
+
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -54,37 +55,35 @@ class VerbHomeFragment : DaggerFragment(), VerbAdapter.OnVerbClickListener {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_verb_home, container, false)
+        return inflater.inflate(R.layout.fragment_adjective_home, container, false)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
+
         initViewModels()
         initRecyclerView()
     }
 
     private fun initViewModels() {
-        verbViewModel = ViewModelProvider(this, providerFactory).get(VerbViewModel::class.java)
+        adjectiveViewModel = ViewModelProvider(this, providerFactory).get(AdjectiveViewModel::class.java)
     }
 
     private fun initRecyclerView() {
-        recycler_verbhome.layoutManager = LinearLayoutManager(context)
-
-        recycler_verbhome.adapter = verbAdapter
+        recycler_adjhome.layoutManager = LinearLayoutManager(context)
+        recycler_adjhome.adapter = adapter
 
         var itemTouchHelper = ItemTouchHelper(
-            VerbSwipeToDeleteHelper(
+            AdjectiveSwipeToDeleteHelper(
                 this,
-                verbViewModel,
-                verbAdapter, 0, ItemTouchHelper.RIGHT
+                adjectiveViewModel,
+                adapter, 0, ItemTouchHelper.RIGHT
             )
         )
 
-        itemTouchHelper.attachToRecyclerView(recycler_verbhome)
-        verbAdapter.notifyDataSetChanged()
-
+        itemTouchHelper.attachToRecyclerView(recycler_adjhome)
+        adapter.notifyDataSetChanged()
     }
-
 
     companion object {
         /**
@@ -93,12 +92,13 @@ class VerbHomeFragment : DaggerFragment(), VerbAdapter.OnVerbClickListener {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment VerbListFragment.
+         * @return A new instance of fragment AdjectiveListFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            VerbHomeFragment().apply {
+            AdjectiveHomeFragment()
+                .apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
@@ -106,7 +106,7 @@ class VerbHomeFragment : DaggerFragment(), VerbAdapter.OnVerbClickListener {
             }
     }
 
-    override fun onVerbClick(id: Int) {
+    override fun onAdjectiveClick(id: Int) {
 
     }
 }
