+++ app/src/main/java/com/teamzmron/selfstudyapp/ui/Fragments/noun/NounAddFragment.kt
@@ -1,20 +1,20 @@
-package com.teamzmron.selfstudyapp.Fragments.Adjective
+package com.teamzmron.selfstudyapp.ui.Fragments.noun
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.EditText
-import android.widget.RadioButton
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.teamzmron.selfstudyapp.Helper.Utils
 import com.teamzmron.selfstudyapp.R
-import com.teamzmron.selfstudyapp.Room.Entity.Adjective
-import com.teamzmron.selfstudyapp.ViewModel.AdjectiveViewModel
+import com.teamzmron.selfstudyapp.Room.Entity.Noun
+import com.teamzmron.selfstudyapp.ViewModel.NounViewModel
 import dagger.android.support.DaggerFragment
-import kotlinx.android.synthetic.main.fragment_adjective.*
+import kotlinx.android.synthetic.main.fragment_noun.*
+
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -23,15 +23,15 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [AdjectiveFragment.newInstance] factory method to
+ * Use the [NounFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class AdjectiveFragment : DaggerFragment() {
+class NounFragment : DaggerFragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
 
-    private lateinit var adjectiveViewModel: AdjectiveViewModel
+    private lateinit var nounViewModel: NounViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -46,17 +46,16 @@ class AdjectiveFragment : DaggerFragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_adjective, container, false)
+        return inflater.inflate(R.layout.fragment_noun, container, false)
     }
+
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
+        nounViewModel = ViewModelProvider(this).get(NounViewModel::class.java)
 
-        adjectiveViewModel = ViewModelProvider(this).get(AdjectiveViewModel::class.java)
 
-        radio_i.isChecked = true
-
-        button_adj_save.setOnClickListener {
+        button_noun_save.setOnClickListener{
             if (checkTextField()) {
                 saveInputFromLayout()
             } else {
@@ -64,48 +63,35 @@ class AdjectiveFragment : DaggerFragment() {
             }
         }
 
-        button_adj_clear.setOnClickListener {
+        button_noun_clear.setOnClickListener {
             clearTextFields()
         }
-
     }
+
 
     private fun saveInputFromLayout() {
-        val adjType = if (getAdjType().contains("な")) "na" else "i"
-        val adjective = Adjective(
-            adjType = adjType,
-            adjNegative = editText_adj_negative.text.toString(),
-            adjPast = editText_adj_past.text.toString(),
-            adjPastNegative = editText_adj_pastnegative.text.toString(),
-            englishWord = editText_adj_englishWord.text.toString(),
-            japaneseWord = editText_adj_japaneseWord.text.toString(),
-            hiraganaForm = editText_adj_hirakata.text.toString(),
-            kanjiForm = editText_adj_kanji.text.toString(),
-            currentTimestamp = Utils.getTimeStamp()
+        val noun = Noun(
+            japanese = editText_noun_japanese.text.toString(),
+            english = editText_noun_english.text.toString(),
+            hiragana = editText_noun_hiragana.text.toString(),
+            kanji = editText_noun_kanji.text.toString(),
+            timestamp = Utils.getTimeStamp()
         )
+        nounViewModel.saveToDB(noun).observe(viewLifecycleOwner, androidx.lifecycle.Observer {
+            if (it > 0) {
+                clearTextFields()
+                Toast.makeText(context!!, "Successfully added!", Toast.LENGTH_LONG).show()
+            }
+        })
 
-        adjectiveViewModel.saveToDB(adjective)
-            .observe(viewLifecycleOwner, androidx.lifecycle.Observer {
-                if (it > 0) {
-                    clearTextFields()
-                    Toast.makeText(context!!, "Successfully added!", Toast.LENGTH_LONG).show()
-                }
-            })
-    }
-
-    private fun getAdjType(): String {
-        return radioGroup_adj.findViewById<RadioButton>(radioGroup_adj.checkedRadioButtonId).text.toString()
     }
 
     private fun getAllTextFields(): ArrayList<EditText> {
         return arrayListOf(
-            editText_adj_negative,
-            editText_adj_past,
-            editText_adj_pastnegative,
-            editText_adj_englishWord,
-            editText_adj_japaneseWord,
-            editText_adj_hirakata,
-            editText_adj_kanji
+            editText_noun_japanese,
+                    editText_noun_english,
+                    editText_noun_hiragana,
+                    editText_noun_kanji
         )
     }
 
@@ -130,7 +116,6 @@ class AdjectiveFragment : DaggerFragment() {
         super.onDestroyView()
         clearTextFields()
     }
-
     companion object {
         /**
          * Use this factory method to create a new instance of
@@ -138,12 +123,12 @@ class AdjectiveFragment : DaggerFragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment AdjectiveFragment.
+         * @return A new instance of fragment NounFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            AdjectiveFragment().apply {
+            NounFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
