+++ app/src/main/java/com/teamzmron/selfstudyapp/ui/Fragments/verb/VerbAddFragment.kt
@@ -1,19 +1,20 @@
-package com.teamzmron.selfstudyapp.Fragments.noun
+package com.teamzmron.selfstudyapp.ui.Fragments.verb
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.EditText
+import android.widget.RadioButton
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.teamzmron.selfstudyapp.Helper.Utils
 import com.teamzmron.selfstudyapp.R
-import com.teamzmron.selfstudyapp.Room.Entity.Noun
-import com.teamzmron.selfstudyapp.ViewModel.NounViewModel
+import com.teamzmron.selfstudyapp.Room.Entity.Verb
+import com.teamzmron.selfstudyapp.ViewModel.VerbViewModel
 import dagger.android.support.DaggerFragment
-import kotlinx.android.synthetic.main.fragment_noun.*
+import kotlinx.android.synthetic.main.fragment_verb.*
 
 
 // TODO: Rename parameter arguments, choose names that match
@@ -23,15 +24,15 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [NounFragment.newInstance] factory method to
+ * Use the [VerbFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class NounFragment : DaggerFragment() {
+class VerbFragment : DaggerFragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
 
-    private lateinit var nounViewModel: NounViewModel
+    private lateinit var verbViewModel: VerbViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -46,52 +47,67 @@ class NounFragment : DaggerFragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_noun, container, false)
+        return inflater.inflate(R.layout.fragment_verb, container, false)
     }
-
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        nounViewModel = ViewModelProvider(this).get(NounViewModel::class.java)
 
+        verbViewModel = ViewModelProvider(this).get(VerbViewModel::class.java)
 
-        button_noun_save.setOnClickListener{
+        radioButton_verb_u.isChecked = true
+
+        button_verb_save.setOnClickListener {
             if (checkTextField()) {
-                saveInputFromLayout()
+                saveInputFromlayout()
             } else {
                 Toast.makeText(context!!, "All fields are required!", Toast.LENGTH_SHORT).show()
             }
+
         }
 
-        button_noun_clear.setOnClickListener {
+        button_verb_clear.setOnClickListener {
             clearTextFields()
         }
     }
 
-
-    private fun saveInputFromLayout() {
-        val noun = Noun(
-            japanese = editText_noun_japanese.text.toString(),
-            english = editText_noun_english.text.toString(),
-            hiragana = editText_noun_hiragana.text.toString(),
-            kanji = editText_noun_kanji.text.toString(),
-            timestamp = Utils.getTimeStamp()
+    private fun saveInputFromlayout() {
+        val verbType = if (getVerbType().contains("")) "u" else "ru"
+        val verb = Verb(
+            verbType = verbType,
+            masu = editText_verb_masu.text.toString(),
+            masuPast = editText_verb_masupast.text.toString(),
+            masuNegative = editText_verb_masunegative.text.toString(),
+            masuPastNegative = editText_verb_masupastnegative.text.toString(),
+            englishWord = editText_verb_englishWord.text.toString(),
+            japaneseWord = editText_verb_japaneseWord.text.toString(),
+            hiraganaForm = editText_verb_hirakata.text.toString(),
+            kanjiForm = editText_verb_kanji.text.toString(),
+            currentTimestamp = Utils.getTimeStamp()
         )
-        nounViewModel.saveToDB(noun).observe(viewLifecycleOwner, androidx.lifecycle.Observer {
+        verbViewModel.saveToDB(verb).observe(viewLifecycleOwner, androidx.lifecycle.Observer {
             if (it > 0) {
                 clearTextFields()
                 Toast.makeText(context!!, "Successfully added!", Toast.LENGTH_LONG).show()
             }
         })
+    }
 
+    private fun getVerbType(): String {
+        return radioGroup_verb.findViewById<RadioButton>(radioGroup_verb.checkedRadioButtonId)
+            .text.toString()
     }
 
     private fun getAllTextFields(): ArrayList<EditText> {
         return arrayListOf(
-            editText_noun_japanese,
-                    editText_noun_english,
-                    editText_noun_hiragana,
-                    editText_noun_kanji
+            editText_verb_masu,
+            editText_verb_masupast,
+            editText_verb_masunegative,
+            editText_verb_masupastnegative,
+            editText_verb_englishWord,
+            editText_verb_japaneseWord,
+            editText_verb_hirakata,
+            editText_verb_kanji
         )
     }
 
@@ -116,6 +132,7 @@ class NounFragment : DaggerFragment() {
         super.onDestroyView()
         clearTextFields()
     }
+
     companion object {
         /**
          * Use this factory method to create a new instance of
@@ -123,12 +140,12 @@ class NounFragment : DaggerFragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment NounFragment.
+         * @return A new instance of fragment VerbFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            NounFragment().apply {
+            VerbFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
