+++ app/src/main/res/layout/fragment_add_word.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:background="@android:color/white"
     android:orientation="vertical"
-    tools:context=".Fragments.AddWord">
+    tools:context=".ui.Fragments.AddWord">
 
 
     <com.google.android.material.tabs.TabLayout
