+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".Fragments.Home">
+    tools:context=".ui.Fragments.Home">
 
     <com.google.android.material.tabs.TabLayout
         android:layout_width="match_parent"
