+++ app/src/main/res/layout/fragment_noun_home.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".Fragments.noun.NounHomeFragment">
+    tools:context=".ui.Fragments.noun.NounHomeFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_nounhome"
