+++ app/src/main/res/layout/fragment_verb_home.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragments.verb.VerbHomeFragment">
+    tools:context=".ui.Fragments.verb.VerbHomeFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <androidx.recyclerview.widget.RecyclerView
