+++ app/src/main/res/layout/quiz_settings_fragment.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@android:color/white"
-    tools:context=".Fragments.QuizSettings">
+    tools:context=".ui.Fragments.QuizSettings">
 
     <ToggleButton
         android:layout_width="match_parent"
