+++ app/src/main/res/menu/main.xml
@@ -2,17 +2,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
-   <item
-       android:id="@+id/home"
-       android:icon="@drawable/ic_home"
-       android:title="Home"
-       app:showAsAction="always"
-       />
-    <item android:id="@+id/add"
-        android:icon="@drawable/ic_add_circle"
-        android:title="Add Vocabulary"
-        app:showAsAction="always"
-        />
     <item
         app:showAsAction="never"
         android:id="@+id/quiz"
